+++ packages/ui/src/components/atoms/Icon.tsx
@@ -1,14 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
 
-interface IconStyledProps extends TIcon {
-  $wght: number
-  $grad: number
-  $opsz: number
-}
-
-const IconStyled = styled.span<IconStyledProps>`
-  font-variation-settings: 'FILL' ${({ fill }) => fill},
+const IconStyled = styled.span<TIcon>`
+  font-variation-settings: 'FILL' ${({ $fill }) => $fill},
     'wght' ${({ $wght }) => $wght}, 'GRAD' ${({ $grad }) => $grad},
     'opsz' ${({ $opsz }) => $opsz};
   color: ${({ color }) => color};
@@ -21,34 +15,37 @@ const IconStyled = styled.span<IconStyledProps>`
 `
 export type TIcon = React.HTMLAttributes<HTMLSpanElement> & {
   name: string
-  fill?: number
+  $fill?: number
   $wght?: number
   $grad?: number
   $opsz?: number
   className?: string
   color?: string
   cursor?: string
+  onClick?: () => void
 }
 
 export const Icon: React.FC<TIcon> = ({
   className = '',
-  fill = 1,
+  $fill = 0,
   $wght = 400,
   $grad = 0,
   $opsz = 48,
   name = '',
   color = '',
+  onClick = () => { },
   ...rest
 }) => (
   <IconStyled
     className={`material-symbols-outlined ${className}`}
-    fill={fill}
+    $fill={$fill}
     $wght={$wght}
     $grad={$grad}
     color={color}
     $opsz={$opsz}
-    {...rest}
     name={name}
+    onClick={onClick}
+    {...rest}
   >
     {name}
   </IconStyled>